// fe/src/lib/api/generated/model/availabilityModeScope.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type AvailabilityModeScope = typeof AvailabilityModeScope[keyof typeof AvailabilityModeScope];


export const AvailabilityModeScope = {
  ALL: 'ALL',
  IN_PERSON: 'IN_PERSON',
  ONLINE: 'ONLINE',
} as const;