// fe/src/lib/api/generated/model/weeklyWindowRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { AvailabilityModeScope } from './availabilityModeScope';
import type { Weekday } from './weekday';

export interface WeeklyWindowRequest {
  end_time: string;
  mode_scope: AvailabilityModeScope;
  start_time: string;
  weekday: Weekday;
}